// web/src/form/questions/response/__generated__/FileQuestionResponseQuery.graphql.ts
export type FileQuestionResponseQuery$variables = {
  formInstanceID: string;
  questionID: string;
};
export type FileQuestionResponseQuery$data = {
  readonly questionResponses: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly __typename: "QuestionResponse";
        readonly id: string;
        readonly value: string;
      } | null | undefined;
    } | null | undefined> | null | undefined;
  };
};
export type FileQuestionResponseQuery = {
  response: FileQuestionResponseQuery$data;
  variables: FileQuestionResponseQuery$variables;
};